b2bxLLM/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
openai
slack_sdk
pytest
langchain
beautifulsoup4
atlassian-python-api
tiktoken